fix: LoadingButton error handling
This commit fixes a bug in the LoadingButton's error-handling. Previously,
if the callback prop throws an error, the `setButtonLoading(false)` call
does not run. This leaves the loading button to continue to spin, even tho
the API response has already been received.

This commit fixes this bug by catching and rethrowing errors in the
callback function, but making sure to set the loading state to be false
first before the error is rethrown.