Feat: empty folders, add resource category (#384)
* Feat: add option to skip adding files

* Fix: reorder page functionality

* Fix: loading and redirection bugs

* Feat: allow empty folders/subfolders

* Fix: prevent immediate redirection on creation of new folder

* Feat: extend folderNamingModal to allow for different types of folders

* Feat: add button for creation of new resource categories in resources

* Feat: allow empty resource folders

* Refactor: use react-query for categorypages

* Refactor: use react-query to retrieve resource category information

* Fix: variable name after rebasing

* Fix: remove timeout

* Fix: hide empty subfolders in EditNavBar

* fix: LoadingButton error handling

This commit fixes a bug in the LoadingButton's error-handling. Previously,
if the callback prop throws an error, the `setButtonLoading(false)` call
does not run. This leaves the loading button to continue to spin, even tho
the API response has already been received.

This commit fixes this bug by catching and rethrowing errors in the
callback function, but making sure to set the loading state to be false
first before the error is rethrown.

Co-authored-by: kwajiehao <jie_hao_kwa@alumni.brown.edu>