feat: replace js-yaml library with yaml library
This commit replaces all use of the npm `js-yaml` library with the
npm `yaml` library.

The main reason for replacing `js-yaml` is because it is unable to
handle certain characters, and throws a YAMLException error. An example
string which it is unable to handle is: `Stairwells, walkways and fresh
food – making today’s cities healthier cities', or in unicode,
`Stairwells, walkways and fresh food â making todayâs cities healthier cities`

However, the `yaml` library is able to handle such strings. Although
it does not have as many weekly downloads as the `js-yaml` library
(27 million weekly downloads), it is still widely used (11 million
weekly downloads) and actively maintained. It also has an easy interface
to use, and substitutes our existing usage of the js-yaml library
without any problems.