fix: improves behavior of MoveDropdown modal
This commit pushes several changes: Firstly, UX of titles and buttons have been improved, including the display of a loading icon while query categories are being retrieved. Secondly, the bug identified earlier where the user would be redirected to the link after returning from the subfolder has been fixed. This means that the user is able to navigate between the subfolder layer and the folders layer, and return to the general dropdown menu if they would like. This fix also fixes the bug behaviors identified by Alex and Jie Hao where dropdown modals are not closed when a user clicks elsewhere on a page, and where queryFolderName is not cleared when a user navigates away from the dropdown modal