Feat/page dropdown api (#388)
* fix: refactor PageSettingsModal to use getEditPageData

* fix: removes getPage from api.js file

* fix: disables Saving PageSettingsModal when no changes

* feat: creates rename and create API calls

* feat: refactors PageSettingsModals saveHandler to use new API calls

* fix: refactor to query for EditPage details once dropdown toggle is activated

* fix: modifies updatePageData to return null when successful

* fix: add delete functionality to FolderPages

* fix: add delete functionality to unlinked pages

* fix: removes SettingsToggle from CollectionPagesSection

* fix: refactors ComponentSettingsModal to remove unused page setting code

* fix: refactor validation for resourceSettings

* fix: refactors ComponentSettingsModal to remove type === 'resources' checking

* fix: fix code bugs and removes unused declarations in  PageSettingsModal

* fix: generateResourceFileName to slugify lower

* fix: refactor ComponentSettingsModal

This commit accomplishes several changes: 1) changes ComponentSettingModal to take pageData as input from CollectionPagesSection and removes the current API call to retrieve page data 2) changes ComponentSettingModal to use useMutation for updating the resource page settings (used for creating new page, renaming a page and changing the url of a page) 3) removes dropdown for selecting page resource category during page resource creation, so that the behavior of creating resources pages will be similar to that of folder pages, where a page can only be created after a user navigates to that location where the new file should be created

* fix: clean up PageSettingsModal

* feat: adds Move To dropdown functionality to folders

* fix: tidy up Folders

* feat: refactor CollectionPagesSection

This commit refactors CollectionPagesSection with the following key changes: 1) Adding Move To functionality to Unlinked pages and restoring Move To functionality to resources 2) cleaning up  variables passed to ComponentSettingsModal since ComponentSettingsModal has been refactored 3) removing unused functionality for retrieving allCategories and improving display of pages prior to completion of loading

* fix: minor tweak to resourceFileName generation

* feat: refactor Overview Card

This commit removes the moveFile API call in Overview Card as it has been replaced by the moveHandler functionality in CollectionPagesSection. This commit also adds Move To functionality for unlinked pages as well as Resources. The Move To functionality for unlinked pages supports moving to both folders and subfolders, but the visual appearance will be improved in a later PR.

* feat: refactor FolderContent

This commit adds Move To functionality for folder and subfolder pages to other folders, other subfolders and to unlinked pages. There are some bugs with the behavior of the back button. The back button and the visual appearance will be improved in a later PR.

* fix: fix MenuDropdown handler behavior when items have nested buttons

* fix: improves MenuDropdown css appearance

* fix: add infoGrey

* fix: add moveFile API, to be cleaned up

* fix: rebase typo with missing } in api.js

* fix: return error for resources

* fix: disables handleBlur

* fix: clean up after settled  mutations

* fix: disables handleBlur

* fix: removes unlinked pages from OverviewCard

* fix: removes filter for folder dropdown

* fix: simplifying api endpoints

* fix: returns null from getCategories when query has not returned

* fix: adds variable to disable blurring, used when fileModalDropdown navigates between folders and subfolders

* fix: changes opacity of dropdownItemChildButton background to 0%

* fix: creates componentTitleLink

* fix: improves behavior of MoveDropdown modal

This commit pushes several changes: Firstly, UX of titles and buttons have been improved, including the display of a loading icon while query categories are being retrieved. Secondly, the bug identified earlier where the user would be redirected to the link after returning from the subfolder has been fixed. This means that the user is able to navigate between the subfolder layer and the folders layer, and return to the general dropdown menu if they would like. This fix also fixes the bug behaviors identified by Alex and Jie Hao where dropdown modals are not closed when a user clicks elsewhere on a page, and where queryFolderName is not cleared when a user navigates away from the dropdown modal

* fix: remove encoding for redirectUrl

* fix: removes unused variables

* fix: rebase error with file filtering

* fix: change numItems check to check for null