fix: header navigation for nested images directories
This commit fixes the back navigation for nested images directories.
Previously, we failed to account for the fact that the custom path
provided as part of the route would be URI encoded. As such, we were
splitting on the `/` character instead of the encoded equivalent,
`%2F`. This commit fixes this issue by using the encoded character
instead.