Feat/nested media layout (#441)
* refactor: Images layout into functional component

* feat: update Images layout to use new API endpoint

* feat: update FolderCard component to accommodate nested media

This commit updates the FolderCard component to accept a `linkPath`
prop so that we can pass in a custom URL for the FolderCard component
to link to. This is necessary for images and files which are nested
in at least one layer of directories.

* feat: update Images layout to display nested directories

* feat: modify Images layout to render nested image folders

This commit modifies the frontend routing to create a /sites/:siteName/images/:customPath
path, which will allow users to access nested image folders on the
Images layout.

The Images layout has also been modified to allow navigation to the
parent image folder using the Header component's back button.

* refactor: Images layout to use react-query

* fix: Image routing for folders nested beyond 2nd level

The frontend paths that we use for Images are:
- /sites/:siteName/images
- /sites/:siteName/images/:customPath

The second path, /sites/:siteName/images/:customPath, is responsible
for handling all other nested levels. For example, if the repo path to
the nested image folder is /images/features/2013, then the CMS URL
would be /sites/:siteName/images/features/2013. The problem with this
is that this URL does not actually match the allocated URL, since
there are `/` characters in the route param. Therefore, we need to
modify the `customPath` variables to be encoded URI components so that
the accessed URLs match the frontend route.

* style: update contentStyles.segment span font size

* feat: update Images layout to match figma specifications

Refer to figma design here: https://www.figma.com/file/4qNhGTBMs521pDvCAIh4ON/IsomerCMS?node-id=1025%3A8585

* feat: add buttons for image upload and directory creation

This commit adds the FolderOptionButton components for image upload
and directory creation.

This commit also fixes a bug where the pending image upload state
was not correctly set when refactoring the Image layout from a class
component to a functional component.

* feat: add error-handling

* fix: rebase error

* fix: use correct icon for image upload as per design specifications

* fix: rebase errors

* fix: header navigation for nested images directories

This commit fixes the back navigation for nested images directories.
Previously, we failed to account for the fact that the custom path
provided as part of the route would be URI encoded. As such, we were
splitting on the `/` character instead of the encoded equivalent,
`%2F`. This commit fixes this issue by using the encoded character
instead.

* fix: modify MediaSettingsModal to retrieve nested files

This commit modifies the endpoint for the GET call to retrieve
images from nested directories instead of just the top level
'images' directory by encoding the path of the nested file.

* fix: update image creation process to allow for creation in nested directories

This commit adds an `imageDirectory` attribute to the request body
to specify the creation of a new image in that nested directory

* fix: update MediaSettingsModal to handle nested images

* Fix: rebase errors

* Feat add image subfolder creation, deletion, renaming functionality (#433)

* Feat: add functionality for create image directory button

* Feat: add rename subfolder functionality

* Refactor: use react-query for FolderCard

* Feat: add functionality for deleting image subfolders

* Fix: message for no images in empty subfolders

* fix: rebase errors

* Feat/files tab revamp (#442)

* Refactor: shift react-query constants into constants file

* Nit: change file/image to files/images for consistency

* Refactor: change image layout to general media layout

* Fix: fix delete images for nested subfolders

* Feat: extend Media layout to handle files

* Fix: rename endpoints for files

This commit renames the existing endpoint for files. Previously, we were using the path `/sites/:siteName/files` for the files tab, but this path could not be extended to handle nested subfolders. The path for the files tab has thus been renamed to `/sites/:siteName/documents`.

* Fix: rebase errors

* Nit: update variable name for clarity

* Fix: change subfolder creation to refetch and toast

* Fix: move refetch media data into useQuery

* feat: refetch data when folders are deleted and renamed

* Feat: add breadcrumb for Media layout

* Chore: remove unused Files component

* Refactor/media settings modal (#453)

* Refactor: use react-query for MediaSettingsModal

* Refactor: change MediaSettingsModal to use react-query

This commit also changes the type parameter of MediaSettingsModal to expect `images` and `files` instead of `image` and `file`

* Fix: refetch instead of reload page on update

* Fix: add customPath to propTypes

* Fix: standardize singular types by using slice

* refactor: updates MediaModal to show nested Media (#450)

* refactor: updates MediaModal to functional component

* fix(UX): implements sticky modalFooter, deslugified directories

* fix: updates FormFieldMedia with refactored MediaModal

* fix: standardizes types from images, files, to image, file

* fix: updates propTypes for refactored MediaModal

* fix: updates FormFieldMedia to accept both image, file type

* fix: add functionality to embed files in EditPage

* nit: standardizes if else logic

* nit: removes additional whitespace

* fix(UX): display successToast upon saving EditPage instead of reload

* fix: replaces type 'image', 'file' with 'images', 'files'

* fix: encoding paths with spaces

* fix: change replaceAll to encodeURIComponent

* fix: add errorToast message for MediaModal

* fix: breadcrumb styling

* fix: uses slice instead of conditional check

* fix: update references to MediaSettingsModal

* fix: changes mediaType indicator in EditPage

* fix: typo in mediaTypes

* fix: callback argument in MediaSettingsModal

* fix: switch back to replaceAll instead of encodeURI

Co-authored-by: gweiying <39231249+gweiying@users.noreply.github.com>

Co-authored-by: gweiying <39231249+gweiying@users.noreply.github.com>

* Nit: rename directory to album for images

* Fix: clear folder name after folder creation

Co-authored-by: jiehao <jiehao@open.gov.sg>
Co-authored-by: kwajiehao <jie_hao_kwa@alumni.brown.edu>
Co-authored-by: gweiying <39231249+gweiying@users.noreply.github.com>