Fix images in private repos not displaying in CMS
Make a call to the backend API to retrieve the b64 encoded image blob if
the link to the raw file from GitHub cannot be found. Converts image
blob to blob url to be set as image src in the MediaSettingsModal,
MediaCard and EditPage components. Removes cached image urls upon image
deletion and image update.