No longer generates Image URL for files in Files tab
Add shouldLoad parameter to fetchImageURL method. MediaCard and
MediaSettingsModal sets shouldLoad to false if the type is not image.
Update useQuery key to include repo name so images are cached
independently of the site. Since retrieving blobs are expensive, change
fetchImageURL staleTime to Infinity so refetches only occur when query
is invalidated.