BREAKING CHANGE: use craco instead of react-scripts
Craco stands for create react app configuration override, and it's a
library which does just that. The reason we're introducing craco is
because create-react-app has complicated webpack config which means
that module-alias does not work with create-react-app. The solution
is to use craco to define webpack aliases instead, through the
craco.config.js file.

This commit does the following:
- uninstall module-alias, and remove module aliases from package.json
- install craco
- replace `react-scripts` with `craco` in package.json scripts
- define path aliases in craco config