chore: add file extension to pacify eslint
eslint complained about not being able to find "@utils/toasts"
because of the mismatch of the extensions between the file importing it,
"src/utils/dropdownUtils.js" and the toasts file, which is .jsx.
This can be fixed through configs, but the easier fix is to just add
the .jsx extension since this only happens once in the entire repo.