style: reset eslint sorting order
Revert import sorting order to the one defined by the example provided:
https://github.com/lydell/eslint-plugin-simple-import-sort/blob/main/examples/.eslintrc.js#L74-L91
We will change the import sort order later on once we have agreed on a
convention.

The reason we're doing that is because our import sort order caused
our own styles to be imported after the sgds styles in src/index.jsx,
which caused our font to change. By reverting to the original import
sorting order, this problem is averted.

This commit also introduces a hack, by importing the react-toastify
styles as an unused variable. This is needed because we need the
react toastify styles BEFORE our element styles for the toast styles
to show up properly. However, this is in contrast with our current
import sorting rules (i.e. we cannot put:

import "react-toastify<blah blah>"

before

import elementStyles from "@styles/<blah blah>"
). The problem is also that we cannot ignore a particular import for
sorting (see https://github.com/lydell/eslint-plugin-simple-import-sort/issues/8).

Therefore, we will leave this here as a hack, since react-toastify is the only
library for which we need this exception for.