style: introduce aliasing, sort imports (#511)
* style: wrap switch cases in blocks

* style: move functions to fix no-use-before-define error

* style: set cy and Cypress as globals in eslint

* style: rename file to jsx because it contains JSX syntax

* style: set param to camelCase

* style: set return for arrow function

* style: replace ++

* style: use @jest/globals

* style: remove unused err and runnable

* style: fix settings tests lint errors

* style: import correct lib

* style: rename RouteSelector from js to jsx

* style: enforce consistent return in api.js

* style: remove unnecessary await

* style: use const instead of let

* style: rename LoginContext file to jsx

* style: lint api.js

* style: fix lint errors for utils.js

* style: set default export

* style: fix lint errors in cspUtils.js

* style: remove unused React

* style: fix inconsistent return error

* style: rename leftnavGeneration from js to jsx

* style: use export default in dropdownUtils.js

* style: do not override result

* style: fix object prototype use issues

* style: fix last linting errors

* fix: import validateFrontMatter correctly

* style: ignore implicitly defined cypress global variables in eslint

* style: install module aliases, setup aliases

This commit installs the module-alias package, and establishes
aliases for top-level directories

* BREAKING CHANGE: use craco instead of react-scripts

Craco stands for create react app configuration override, and it's a
library which does just that. The reason we're introducing craco is
because create-react-app has complicated webpack config which means
that module-alias does not work with create-react-app. The solution
is to use craco to define webpack aliases instead, through the
craco.config.js file.

This commit does the following:
- uninstall module-alias, and remove module aliases from package.json
- install craco
- replace `react-scripts` with `craco` in package.json scripts
- define path aliases in craco config

* refactor: replace path imports with aliases

This commmit replaces all path imports with their respective aliases
as defined in the craco config file. There were two exceptions:
- the route selector test file
- the path provided to the url function in src/styles/isomer-cms/pages/Media.module.scss

We will need to find a way to reconcile paths for the jest test
in a subsequent commit.

* style: use the proper prettier plugin for eslint

This commit modifies our eslint config to use the `prettier` plugin
instead of the `plugin:prettier/recommended` plugin as it interferes
with our linting and code formatting process. Refer to the backend PR
https://github.com/isomerpages/isomercms-backend/pull/195 for more
details.

* style: install eslint-import-resolver-alias

This commit installs the eslint-import-resolver-alias plugin so
that eslint can understand the same aliases that we've defined in
the craco config file

* chore: add file extension to pacify eslint

eslint complained about not being able to find "@utils/toasts"
because of the mismatch of the extensions between the file importing it,
"src/utils/dropdownUtils.js" and the toasts file, which is .jsx.
This can be fixed through configs, but the easier fix is to just add
the .jsx extension since this only happens once in the entire repo.

* chore: re-add module aliases

* chore: remove unused tree files and packages

* chore: update config for eslint aliases

* style: define simple-import-sort configs

* style: add no-underscore-dangle rule, consistent with backend eslint

* style: modify simple-import-sort rules

This commit makes two changes:
- Use (.*|$) instead of the provided (\.*|$) regex for matching
suffix characters for groups.
- Introduce a break between react-related external package imports
and non-react external package imports (from ["^react", "^@?\\w"] to
["^react(.*|$)"], ["^@?\\w"])

Regarding the first change, previously, we were unable to match
strings containing sub-directories properly, such as
"@styles/isomer-cms/Elements.module.scss". This new change allows us to
do so.

* style: autofix import sorting

This commit uses eslint to autofix import sorting using the following
command: npx eslint --ext .js --ext .jsx --plugin eslint-plugin-simple-import-sort . --fix

Comments such as "// Import components" or "// Import contexts" were
also removed, as they are now redundant.

* style: add webpack aliases for VS Code intellisense

This commit adds the webpack aliases so that VS Code can detect
them. We referred to the following documentation from the VS
Code website: https://code.visualstudio.com/docs/languages/jsconfig#_using-webpack-aliases

* style: add aliases for jest

This commit adds the path aliases for jest so that our tests
can run. We used the jest moduleNameMapper to achieve this effect,
referring to this article: https://medium.com/@justintulk/solve-module-import-aliasing-for-webpack-jest-and-vscode-74007ce4adc9

* style: run prettier

* chore: fix merge conflict errors

* style: rejig style files import sorting

* style: add aliases for cypress webpack config

This commit adds an index.js file which adds our aliases to cypress
during the file preprocessing stage using the @cypress/webpack-preprocessor
package.

It also removes unnecessary code from the test files.

* style: remove padding on bottom of page

* chore: uncomment out failing Jest mocks

Currently, our route selector tests are failing because some
of our mocks reference Images and Files layouts which we have
already deleted.

An issue has been filed to track this:
https://github.com/isomerpages/isomercms-frontend/issues/513

* style: reset eslint sorting order

Revert import sorting order to the one defined by the example provided:
https://github.com/lydell/eslint-plugin-simple-import-sort/blob/main/examples/.eslintrc.js#L74-L91
We will change the import sort order later on once we have agreed on a
convention.

The reason we're doing that is because our import sort order caused
our own styles to be imported after the sgds styles in src/index.jsx,
which caused our font to change. By reverting to the original import
sorting order, this problem is averted.

This commit also introduces a hack, by importing the react-toastify
styles as an unused variable. This is needed because we need the
react toastify styles BEFORE our element styles for the toast styles
to show up properly. However, this is in contrast with our current
import sorting rules (i.e. we cannot put:

import "react-toastify<blah blah>"

before

import elementStyles from "@styles/<blah blah>"
). The problem is also that we cannot ignore a particular import for
sorting (see https://github.com/lydell/eslint-plugin-simple-import-sort/issues/8).

Therefore, we will leave this here as a hack, since react-toastify is the only
library for which we need this exception for.

Co-authored-by: Preston Lim <prestonlimlianjie@gmail.com>