Lint/update husky linter (#518)
* lint: install lint-staged package
This commit installs the lint-staged package, which will help us
to automatically fix linting and formatting errors as commits are
being made. This install comes with a husky pre-commit hook.

* lint: add eslint and prettier autofix commands to lint-staged

* lint: install eslint-plugin-only-warn
This commit installs the only-warn plugin for eslint, which converts
all errors into warnings. The reason for doing this is because we
wish to introduce automatic linting and formatting into our repo,
but there remain some linting errors that require manual attention
(for example, we currently have some files which violate the
max-classes-per-file rule) which we wish to tackle at a later date.

Using this plugin allows us to use lint-staged to automatically fix
linting and formatting errors in the pre-commit stage, while also allowing
us to ignore errors emitted by eslint, since these are now treated as
warnings.