fix: remove duplicate html id tag for DeleteWarningModal
There currently exists two id attributes for the `DeleteWarningModal`
component: "modal-delete" and "delete-confirmation". Each id is used
in one of our e2e tests. However, since "delete-confirmation" is specified
after "modal-delete", "modal-delete" is overwritten and so the test
relying on "#modal-delete" fails.

This commit fixes this bug by removing the extraneous "delete-confirmation"
id and standardizing e2e tests to use the "modal-delete" id instead.
"modal-delete" was chosen over "delete-confirmation" because the id
attribute for the cancel button next to the delete button is "modal-cancel".