test: fix flaky delete test for EditPage
Previously, the "Edit page should allow user to delete page" test
failed because Cypress was unable to click on the delete button
in the DeleteWarningModal. We suspect that this is because the
modal takes some time to appear in the DOM.

To resolve this problem, we wait for the DeleteWarningModal to exist
by using a "should" assertion before attempting to click on it.