test: reset e2e-test-repo before running cypress tests
This commit uses each individual user's personal access token to
reset the e2e-test-repo to a referenced commit hash so that we don't
need to do this before running our e2e tests locally.

As it stands, the repo is reset once before all e2e tests are run.
This should be sufficient since our tests should not be dependent
on side effects to pass (i.e. we shouldn't rely on the outputs of
one of our tests for a subsequent test to pass).