test: fix detached breadcrumb element problem when moving files
This commit uses the cypress-pipe library (on recommendation from the
Cypress team) to help to click on the correct DOM element when we're
trying to move a file from one directory to another.

Previously, Cypress would complain that the breadcrumb element was
detached from the DOM, despite already asserting for visibility. This
is a case where the app is slow to react and Cypress was fast to act:
see the following blog post from the Cypress team which addresses the
issue: https://www.cypress.io/blog/2019/01/22/when-can-the-test-click/
The problem is probably that the DOM element has been created but
the event listeners have not been set up.

cypress-pipe solves this problem by continuously retrying until
“the assertions that follow pass”.