chore: update encode decode method for thirdnavtitle & filename
This commit updates the decoding and encoding of third-nav-title and filename, which affects the validation, display and endpoints for both third-nav-title and filenames. After discussion with Alex, we've decided to go with the following:
filename: Allows any special characters, any random capitalization of letters, uses raw output of encodeURI & decodeURI
thirdnavtitle: Allows special characters and any random capitalization of letters APART from first letter after each dash, on the backend it uses the output of decodeURI, and splits by dashes and capitalizes each of the split strings. This was done so that it would be compatible with the current migrated repos.

See the design document for more details about this