build(package): installed many dependencies and plugins
craco is needed to process mjs properly, as we need to run it through a preprocessor. this then
leads to storybook-preset-craco being required for storybook to process them the same way as
storybook uses cra initially.

additionally, react 17 has a react/jsx-runtime, which caused issues with the existing version of react (16.xx), which did not have this. This required an additional upgrade of recat from 16.xx to 17 and a further configuration change in tsconfig: