refactor(formfieldmedia): changed ThemeProvider to be story-based
the theme provider (rightfully) alters certain aspects of components. if we set it to be applied
throughout (which should be the case once we are fully shifted over), this will lead to components
in the story not matching up against the actual components on prod. hence, we have to go for the
more verbose case and let the theming be per story