build(react-dom): pins react-dom to 17.0.2
react is on 17.xx, which uses a different transform. we need to pin react-dom to the same version
(or higher) to maintain compatibility i think. without this, there was a runtime error on storybook,
where it built successfully but failed to display the webpage, showing "Uncaught undefined" in the
console