test(cypress): increased wait times
some of our tests are flaky because of time outs. this is not ideal because tests should be easily
reproducible. hence, the timeout has been increased to 10s. this change is probably only going to
affect prod pushes so dev impact should hopefully be minimal once integation tests are up