feat: add story for formFieldMedia (#759)
* test(formfieldmedia): add initial story for formfieldmedia

* refactor(formfieldmedia): updated stories to fit styling

* refactor(formfieldmedia): changed ThemeProvider to be story-based

the theme provider (rightfully) alters certain aspects of components. if we set it to be applied
throughout (which should be the case once we are fully shifted over), this will lead to components
in the story not matching up against the actual components on prod. hence, we have to go for the
more verbose case and let the theming be per story

* chore(formfieldmedia): updated stories

* feat(formfieldmedia): ports over to typescript and design system

* docs(formfieldmedia): add v2 stories

* feat(formfieldmedia): add v2 of form field media that uses the design system components

* refactor(formfieldmedia): minor refactor and clean up of stories

* chore(chromatic): test commit for chromatic integration wtih github