feat(design-system): adds storybook/chromatic and ogp design system  (#757)
* build(storybook): scaffolded storybook

* build(typescript): installed typescript together with a tsconfig

* build(types): add types for react router dom

* chore(.eslint): update eslint rules to work with ts

* build(design system): installed ogp design system and peer deps

* chore(storybook): update storybook to use ogp design system theme

* build(package): installed many dependencies and plugins

craco is needed to process mjs properly, as we need to run it through a preprocessor. this then
leads to storybook-preset-craco being required for storybook to process them the same way as
storybook uses cra initially.

additionally, react 17 has a react/jsx-runtime, which caused issues with the existing version of react (16.xx), which did not have this. This required an additional upgrade of recat from 16.xx to 17 and a further configuration change in tsconfig:

* build(craco config): adds a configuration rule for mjs so that the default export problem is fixed

* build(tsconfig): upgrades tsconfig to use new jsx parsing

* build(storybook): updates storybook config to use same webpack config as craco

* chore(eslint): update eslint for react 17 js runtime

* build(react-dom): pins react-dom to 17.0.2

react is on 17.xx, which uses a different transform. we need to pin react-dom to the same version
(or higher) to maintain compatibility i think. without this, there was a runtime error on storybook,
where it built successfully but failed to display the webpage, showing "Uncaught undefined" in the
console

* build(eslint): update eslint-react-plugin to latest version for new plugin rules

* build(chromatic): adds chromatic

* chore(gitginore): update gitignore to ignore chromatic build log

* chore(eslint): update eslint rules

* ci(chromatic): adds a chromatic workflow to ci

* fix(chromatic): manually pin node version in chromatic workflow

we need to manually pin node version because libsass fails to build on node 16+, which is the
default.

* build(package): update to add missing emotion/styled dependency

* feat: add story for formFieldMedia (#759)

* test(formfieldmedia): add initial story for formfieldmedia

* refactor(formfieldmedia): updated stories to fit styling

* refactor(formfieldmedia): changed ThemeProvider to be story-based

the theme provider (rightfully) alters certain aspects of components. if we set it to be applied
throughout (which should be the case once we are fully shifted over), this will lead to components
in the story not matching up against the actual components on prod. hence, we have to go for the
more verbose case and let the theming be per story

* chore(formfieldmedia): updated stories

* feat(formfieldmedia): ports over to typescript and design system

* docs(formfieldmedia): add v2 stories

* feat(formfieldmedia): add v2 of form field media that uses the design system components

* refactor(formfieldmedia): minor refactor and clean up of stories

* chore(chromatic): test commit for chromatic integration wtih github