chore(waitfordom): remove waitfordom and related files
this is because the wait for dom file runs into errors with pre-processing using webpack (see our
cypress config, which loads and uses webpack). this means that the file is currently incompatible
with ts (however, it does work if js). as this file is only used in medias, the file is removed and
the call sites will be updated accordingly