build(index): remove bootstrap
i double checked this by gg to the templates repo and searching for css impotrs both in the
`next-gen` branch and in `staging`. no bootstrap import was spotted so this should be a safe removal
assuming we aren't depending on it for styling for template components. for cms components, we will
use design system for styling instead.