fix(permalinks): default permalink for create (#1843)
## Problem

We have quite a number of duplicate permalinks
this does not fix the problem completely, but it does allow for some sort of preventive measure to prevents accidental dups. 

we are not sure if of the number of agencies that use dup permalinks for the sake of playground, and it would be hard to enforce this until we have the functionality for playground out. more context [here](https://opengovproducts.slack.com/archives/CK68JNFHR/p1710317487152189)


## Solution

sane default when creating a page.

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [ ] Yes - this PR contains breaking changes
  - Details ...
- [X] No - this PR is backwards compatible with ALL of the following feature flags in this [doc](https://www.notion.so/opengov/Existing-feature-flags-518ad2cdc325420893a105e88c432be5)


## Tests

follow video

https://github.com/isomerpages/isomercms-frontend/assets/42832651/3c3657dc-b9a1-4a40-9e49-07d71c32fd0a

- [ ] create a page, note that you should be unable to modify the permalink 
- [ ] edit page modal's functionality remains the same, we dont still allow them to change to whatever.